UF-14556: Acted on PR comments
- Adjusted code according to PR comments
- Moved intric information model to better named package
- Moved logic for updating intric model with extended data to own
decorator classes
- Adjusted Intric followup question api to enable possibility to send
ids of information files to base answer on
