+++ src/main/java/se/sundsvall/selfserviceai/integration/intric/model/AskAssistant.java
@@ -1,9 +1,11 @@
 package se.sundsvall.selfserviceai.integration.intric.model;
 
+import java.util.List;
 import lombok.Builder;
 
 @Builder(setterPrefix = "with")
 public record AskAssistant(
 
-	String question) {
+	String question,
+	List<String> files) {
 }
