+++ src/test/java/se/sundsvall/selfserviceai/integration/intric/IntricIntegrationTest.java
@@ -5,6 +5,7 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import java.util.List;
 import java.util.UUID;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -36,20 +37,20 @@ class IntricIntegrationTest {
 	 */
 	@Test
 	void askAssistant_1() {
-		var assistantId = "assistantId";
-		var input = "input";
-		var sessionId = "2d357dcf-6180-48de-a9e8-3ad74b757c84";
-		var answer = "answer";
-		var response = AskResponse.builder()
+		final var assistantId = "assistantId";
+		final var input = "input";
+		final var sessionId = "2d357dcf-6180-48de-a9e8-3ad74b757c84";
+		final var answer = "answer";
+		final var response = AskResponse.builder()
 			.withSessionId(UUID.fromString(sessionId))
 			.withQuestion(input)
 			.withAnswer(answer)
 			.build();
 		when(intricClientMock.askAssistant(eq(assistantId), askAssistantCaptor.capture())).thenReturn(response);
 
-		var result = integration.askAssistant(assistantId, input);
+		final var result = integration.askAssistant(assistantId, input);
 
-		var askAssistant = askAssistantCaptor.getValue();
+		final var askAssistant = askAssistantCaptor.getValue();
 		assertThat(askAssistant.question()).isEqualTo(input);
 
 		assertThat(result).isEqualTo(response);
@@ -61,18 +62,16 @@ class IntricIntegrationTest {
 	 */
 	@Test
 	void askAssistant_2() {
-		var assistantId = "assistantId";
-		var input = "input";
+		final var assistantId = "assistantId";
+		final var input = "input";
 		when(intricClientMock.askAssistant(eq(assistantId), askAssistantCaptor.capture())).thenThrow(new RuntimeException("Something went wrong"));
 
-		var result = integration.askAssistant(assistantId, input);
+		final var result = integration.askAssistant(assistantId, input);
 
-		var askAssistant = askAssistantCaptor.getValue();
+		final var askAssistant = askAssistantCaptor.getValue();
 		assertThat(askAssistant.question()).isEqualTo(input);
 
-		assertThat(result.sessionId()).isNull();
-		assertThat(result.answer()).isBlank();
-		assertThat(result.question()).isEqualTo(input);
+		assertThat(result).isNull();
 		verify(intricClientMock).askAssistant(assistantId, askAssistant);
 	}
 
@@ -81,20 +80,23 @@ class IntricIntegrationTest {
 	 */
 	@Test
 	void askFollowUp_1() {
-		var assistantId = "assistantId";
-		var sessionId = "2d357dcf-6180-48de-a9e8-3ad74b757c84";
-		var input = "input";
-		var response = AskResponse.builder()
+		final var assistantId = "assistantId";
+		final var sessionId = "2d357dcf-6180-48de-a9e8-3ad74b757c84";
+		final var input = "input";
+		final var fileReferences = List.of("fileId");
+		final var response = AskResponse.builder()
 			.withSessionId(UUID.fromString(sessionId))
 			.withQuestion(input)
 			.withAnswer("answer")
 			.build();
+
 		when(intricClientMock.askFollowUp(eq(assistantId), eq(sessionId), askAssistantCaptor.capture())).thenReturn(response);
 
-		var result = integration.askFollowUp(assistantId, sessionId, input);
+		final var result = integration.askFollowUp(assistantId, sessionId, input, fileReferences);
 
-		var askAssistant = askAssistantCaptor.getValue();
+		final var askAssistant = askAssistantCaptor.getValue();
 		assertThat(askAssistant.question()).isEqualTo(input);
+		assertThat(askAssistant.files()).isEqualTo(fileReferences);
 
 		assertThat(result).isEqualTo(response);
 		verify(intricClientMock).askFollowUp(assistantId, sessionId, askAssistant);
@@ -105,20 +107,20 @@ class IntricIntegrationTest {
 	 */
 	@Test
 	void askFollowUp_2() {
-		var assistantId = "assistantId";
-		var sessionId = "2d357dcf-6180-48de-a9e8-3ad74b757c84";
-		var input = "input";
+		final var assistantId = "assistantId";
+		final var sessionId = "2d357dcf-6180-48de-a9e8-3ad74b757c84";
+		final var input = "input";
+		final var fileReferences = List.of("fileId");
 
 		when(intricClientMock.askFollowUp(eq(assistantId), eq(sessionId), askAssistantCaptor.capture())).thenThrow(new RuntimeException("Something went wrong"));
 
-		var result = integration.askFollowUp(assistantId, sessionId, input);
+		final var result = integration.askFollowUp(assistantId, sessionId, input, fileReferences);
 
-		var askAssistant = askAssistantCaptor.getValue();
+		final var askAssistant = askAssistantCaptor.getValue();
 		assertThat(askAssistant.question()).isEqualTo(input);
+		assertThat(askAssistant.files()).isEqualTo(fileReferences);
 
-		assertThat(result.answer()).isBlank();
-		assertThat(result.question()).isEqualTo(input);
-		assertThat(result.sessionId()).isEqualTo(UUID.fromString(sessionId));
+		assertThat(result).isNull();
 		verify(intricClientMock).askFollowUp(assistantId, sessionId, askAssistant);
 	}
 
@@ -127,9 +129,9 @@ class IntricIntegrationTest {
 	 */
 	@Test
 	void uploadFile_1() {
-		var multiPartFileMock = Mockito.mock(MultipartFile.class);
-		var id = "2d357dcf-6180-48de-a9e8-3ad74b757c84";
-		var filePublic = FilePublic.builder()
+		final var multiPartFileMock = Mockito.mock(MultipartFile.class);
+		final var id = "2d357dcf-6180-48de-a9e8-3ad74b757c84";
+		final var filePublic = FilePublic.builder()
 			.withId(UUID.fromString(id))
 			.withName("name")
 			.withMimeType("mimeType")
@@ -137,7 +139,7 @@ class IntricIntegrationTest {
 			.build();
 		when(intricClientMock.uploadFile(multiPartFileMock)).thenReturn(filePublic);
 
-		var result = integration.uploadFile(multiPartFileMock);
+		final var result = integration.uploadFile(multiPartFileMock);
 
 		assertThat(result).isEqualTo(UUID.fromString(id));
 
@@ -149,10 +151,10 @@ class IntricIntegrationTest {
 	 */
 	@Test
 	void uploadFile_2() {
-		var multiPartFileMock = Mockito.mock(MultipartFile.class);
+		final var multiPartFileMock = Mockito.mock(MultipartFile.class);
 		when(intricClientMock.uploadFile(multiPartFileMock)).thenThrow(new RuntimeException("Something went wrong"));
 
-		var result = integration.uploadFile(multiPartFileMock);
+		final var result = integration.uploadFile(multiPartFileMock);
 
 		assertThat(result).isNull();
 		verify(intricClientMock).uploadFile(multiPartFileMock);
@@ -163,7 +165,7 @@ class IntricIntegrationTest {
 	 */
 	@Test
 	void deleteFile() {
-		var fileId = "2d357dcf-6180-48de-a9e8-3ad74b757c84";
+		final var fileId = "2d357dcf-6180-48de-a9e8-3ad74b757c84";
 
 		integration.deleteFile(fileId);
