+++ src/test/java/se/sundsvall/selfserviceai/integration/intric/model/filecontent/MeasurementDataTest.java
@@ -1,4 +1,4 @@
-package se.sundsvall.selfserviceai.service.model;
+package se.sundsvall.selfserviceai.integration.intric.model.filecontent;
 
 import static com.google.code.beanmatchers.BeanMatchers.hasValidBeanConstructor;
 import static com.google.code.beanmatchers.BeanMatchers.hasValidBeanEquals;
@@ -34,7 +34,7 @@ class MeasurementDataTest {
 	}
 
 	@Test
-	void testBuilderMethodsForAddress() {
+	void testBuilderMethods() {
 		final var category = "category";
 		final var measurementType = "measurementType";
 		final var timestamp = OffsetDateTime.now();
